Port cbed_frame into utils
